+++ include/swift/SIL/SILBridging.h
@@ -1538,7 +1538,7 @@ struct BridgedCloner {
   bool isValueCloned(BridgedValue v) const;
   void recordClonedInstruction(BridgedInstruction origInst, BridgedInstruction clonedInst) const;
   void recordFoldedValue(BridgedValue orig, BridgedValue mapped) const;
-  BridgedInstruction clone(BridgedInstruction inst);
+  BridgedInstruction clone(BridgedInstruction inst) const;
 };
 
 struct BridgedVerifier {
